requirements.txt
langchain>=0.2.14,<0.4
langchain-community>=0.2.12,<0.4
langchain-gigachat>=0.3.0,<0.6
pydantic>=2.7,<3
python-dotenv>=1.0.1,<2
requests>=2.31,<3

